feat: solve modules/esm
